Implement token with the `enumeration` macro
I'm an idiot. I previously implemented Token by hand as an enum combining both
Piece and Color because piece doesn't fit cleanly into three bits. I kept a
two-element gap between white and black pieces, so it would be easy to
bit-twiddle out the color by checking the fourth bit. If the two components of
the enum were consecutive, doing this would involve extra (slower) work to tease
out the independent color and piece components.

The trivial solution is to alternate white and black pieces in the enum. This
causes the LSB to always be color, and the three next bits to always be the
piece type. All the problems go away, the enum is fully consecutive with no
gaps, and bit twiddling is trivial.